refactor: in-memory files changed
